c300: extract Embed3DTab to own file (#441 2.1)
Workspace.tsx 7568 LOC -> 7362 LOC (-206 LOC). Entry chunk 190.20 KB -> 186.56 KB.
19 tabs extracted + lazy. Cumulative since c262: -40%. typecheck clean, 44/44 tests pass.
DocsPerTopicBar kept inline (used by TopicLabelTab + TopicsTab).
